requirements.txt
mcp[cli]
python-dotenv==1.1.1



httpx==0.28.1
beautifulsoup4==4.13.5